+++ sdk/cognitivelanguage/Azure.AI.Language.Conversations/tests/ConversationAnalysisClientLiveTests.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Net.Sockets;
 using System.Text.Json;
 using System.Threading.Tasks;
 using Azure.Core.TestFramework;
@@ -164,6 +165,26 @@ namespace Azure.AI.Language.Conversations.Tests
 
             // assert - inent target kind
             Assert.AreEqual(TargetProjectKind.QuestionAnswering, topIntent.TargetProjectKind);
+        }
+
+        [RecordedTest]
+        public async Task SupportsAadAuthentication()
+        {
+            ConversationAnalysisClient client = CreateClient<ConversationAnalysisClient>(
+                TestEnvironment.Endpoint,
+                TestEnvironment.Credential,
+                InstrumentClientOptions(
+                    new ConversationAnalysisClientOptions(ServiceVersion)));
+
+            ConversationalTask conversationalTask = new(
+                new ConversationAnalysisOptions(new TextConversationItem("1", "1", "Send an email to Carol about the tomorrow's demo")),
+                new ConversationTaskParameters(TestEnvironment.ProjectName, TestEnvironment.DeploymentName));
+
+            Response response = await client.AnalyzeConversationAsync(conversationalTask.AsRequestContent());
+
+            using JsonDocument json = await JsonDocument.ParseAsync(response.ContentStream);
+            ConversationalTaskResult conversationalTaskResult = ConversationalTaskResult.DeserializeConversationalTaskResult(json.RootElement);
+            Assert.That(conversationalTaskResult.Result.Prediction.TopIntent, Is.EqualTo("Send"));
         }
     }
 }